v0.8.1: release-readiness hardening for watcher, task snapshots, and swarm members.
- sync: fix watcher .git prefix matching, add onError callback, make Close()
  idempotent and flush pending changes, surface walk/fsnotify errors
- mesh: deep-copy ToolResult in task snapshots, harden test
- swarm: explicit MemberInfo field copy with compile-time shape guard
- lint: replace misleading filepath.Clean with //nolint:gosec rationales
- p2p: single OpenProtocolStream for agent/mesh/swarm/sync transports,
  grace-wait for ReliableConn.WriteFrame, fix transport announce test bound
- ci/test: re-enable errcheck/staticcheck/gosec, per-OS diskspace files,
  Windows test scripts, parallelization and timeout hardening

+++ pkg/browser/diskspace_darwin.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build darwin
 
 package browser
 
@@ -10,5 +10,5 @@ func FreeBytes(dir string) (uint64, error) {
 	if err := unix.Statfs(dir, &st); err != nil {
 		return 0, err
 	}
-	return st.Bavail * uint64(st.Bsize), nil
+	return uint64(st.Bavail) * uint64(st.Bsize), nil
 }
